pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jwks-server"
version = "0.1.0"
description = "Educational JWKS server with key expiry and auth endpoint"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pyjwt[crypto]>=2.8.0",
    "cryptography>=42.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
